ui/pyproject.toml
[tool.poetry]
name = "ui"
version = "0.1.0"
description = ""
authors = ["Malik GAruba <garubamalik@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
streamlit = "^1.36.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
